requirements.txt
streamlit==0.73.1
numpy==1.18.1
pandas==1.0.0
psycopg2==2.8.4
python-dotenv==0.12.0
